APM-270 Seperate out sandbox server into seperate files
