+++ specification/components/examples/Patient-John-Smythe.json
@@ -1,6 +1,6 @@
 {
     "resourceType": "Patient",
-    "id": "9000000009",
+    "id": "90000000010",
     "identifier": [
       {
         "system": "https://fhir.nhs.uk/Id/nhs-number",
